backend/requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1
Flask==3.0.3
flask_cors==5.0.0
PyJWT==2.9.0